services/scheduled-consumers/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loom-scheduled-consumers"
version = "0.1.0"
description = "Scheduled consumers for external data ingestion in Loom v2"
authors = [
    {name = "Loom Team", email = "team@loom.dev"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiokafka>=0.12.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=25.0.0",
    "feedparser>=6.0.0",  # For RSS feeds
    "beautifulsoup4>=4.12.0",  # For web scraping
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.0",
    "schedule>=1.2.0",  # For cron-like scheduling
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "S", "C90"]
ignore = ["E501", "S101"]  # Line too long, assert usage

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true